Changed readLength to a constant of size_t.

+++ src/Utilities/INIParser.h
@@ -13,7 +13,7 @@ class INI_EX {
 		return Ares::readBuffer;
 	}
 
-	int buflen() {
+	size_t buflen() {
 		return Ares::readLength;
 	}
